Improve documentation. Increase version in documentation and installer. Move all CefSharp handlers in Handlers folder. Prepare to download.
