+++ docs/404.html
@@ -28,7 +28,7 @@ layout: default
 	  <p>The development site is in <a href="{{site.project_url}}">{{site.github_img}} <span class="username">yoyokits/VideoBrowser</span></a>  </p>
 
 <p align="center">
-   <img src="assets/images/cekli-logo.png"><br>	
+   <img src="/VideoBrowser/assets/images/cekli-logo.png"><br>	
    <strong>Cekli Browser and Youtube Downloader</strong>
 </p>
