+++ docs/Gemfile
@@ -10,6 +10,7 @@ source "https://rubygems.org"
 # gem "jekyll", "~> 4.0.1"
 # This is the default theme for new Jekyll sites. You may change this to anything you like.
 gem "minima", "~> 2.5"
+# gem "minimal-mistakes-jekyll"
 # If you want to use GitHub Pages, remove the "gem "jekyll"" above and
 # uncomment the line below. To upgrade, run `bundle update github-pages`.
 # gem "github-pages", group: :jekyll_plugins
@@ -22,6 +23,8 @@ group :jekyll_plugins do
   gem 'jekyll-seo-tag', '~> 2.6', '>= 2.6.1'
   gem 'jekyll-sitemap'
   gem "jekyll-youtube"
+  gem 'jekyll-include-cache'
+  gem 'jekyll-remote-theme'
 end
 
 # Windows and JRuby does not include zoneinfo files, so bundle the tzinfo-data gem
