+++ docs/_config.yml
@@ -19,7 +19,7 @@
 # in the templates via {{ site.myvariable }}.
 
 title: Cekli Video Browser and Downloader
-version: v0.1.3
+version: v0.1.4
 email: 
 description: >- # this means to ignore newlines until "baseurl:"
   Cekli Video Browser is free a Chromium-based browser optimized for downloading Youtube Videos.
@@ -31,12 +31,15 @@ github_username:  yoyokits/VideoBrowser
 
 # Build settings
 theme: minima
+#theme: minimal-mistakes-jekyll
+
 plugins:
   - jekyll-analytics
   - jekyll-feed
   - jekyll-seo-tag
   - jekyll-sitemap
   - jekyll-youtube
+  - jekyll-include-cache
 
 # Exclude from processing.
 # The following items will not be processed, by default.
@@ -59,9 +62,9 @@ plugins:
 #   - vendor/ruby/
 
 # Custom variable:
-contact: <ul class="social-media-list"><li><a href="https://github.com/yoyokits/VideoBrowser"><svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/minima-social-icons.svg#github"></use></svg> <span class="username">yoyokits/VideoBrowser</span></a></li><li><a href="https://www.twitter.com/nurcahyoyohanes"><svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/minima-social-icons.svg#twitter"></use></svg>  <span class="username">nurcahyoyohanes</span></a></li></ul>
+contact: <ul class="social-media-list"><li><a href="https://github.com/yoyokits/VideoBrowser"><svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/images/minima-social-icons.svg#github"></use></svg> <span class="username">yoyokits/VideoBrowser</span></a></li><li><a href="https://www.twitter.com/nurcahyoyohanes"><svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/images/minima-social-icons.svg#twitter"></use></svg>  <span class="username">nurcahyoyohanes</span></a></li></ul>
 project_url: https://github.com/yoyokits/VideoBrowser
 twitter_url: https://www.twitter.com/nurcahyoyohanes
-twitter_img: <svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/minima-social-icons.svg#twitter"></use></svg>
-github_img: <svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/minima-social-icons.svg#github"></use></svg>
+twitter_img: <svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/images/minima-social-icons.svg#twitter"></use></svg>
+github_img: <svg class="svg-icon"><use xlink:href="/VideoBrowser/assets/images/minima-social-icons.svg#github"></use></svg>
 image_folder: /VideoBrowser/assets/images/
