+++ docs/assets/css/styles.scss
@@ -2,6 +2,7 @@
 ---
 
 @import "{{ site.theme }}";
+//@import "minimal-mistakes";
 
 .logo_container { 
   width: 80px;
