+++ docs/index.markdown
@@ -18,10 +18,10 @@ Bug reports or following the development progress can be reached via:
 
 {{site.contact}}
 
-![Cekli Video Browser Screenshot]({{site.image_folder}}/{{site.version}}/CekliVideoBrowserDownloadOptionsScreenShot.jpg)
+![Cekli Video Browser Screenshot]({{site.image_folder}}{{site.version}}/CekliVideoBrowserDownloadOptionsScreenShot.jpg)
 Cekli Browser screenshot
 
-![Cekli Video Browser Screenshot]({{site.image_folder}}/{{site.version}}/CekliVideoBrowserDownload.gif)
+![Cekli Video Browser Screenshot]({{site.image_folder}}{{site.version}}/CekliVideoBrowserDownload.gif)
 Download youtube with Cekli Browser
 
 
@@ -55,7 +55,7 @@ It has dependencies (usually already installed in your system)  with:
 
 **Usage**
 
-![{{site.title}} as standard browser]({{site.image_folder}}/{{site.version}}/CekliVideoBrowserMainViewScreenShot.jpg)
+![{{site.title}} as standard browser]({{site.image_folder}}{{site.version}}/CekliVideoBrowserMainViewScreenShot.jpg)
 Like any other web browsers, {{site.title}} can be used as a web browser.
 
 If we browse in a youtube video then the download icon will be enabled, after clicking then it will be downloaded in a queue.
@@ -63,21 +63,21 @@ It can downloads several video concurently or multithreading.
 
 > **Warning:** some Video websites not allowing or limiting download number in a time, please make sure of it.
 
-![Doownload Queue]({{site.image_folder}}/{{site.version}}/CekliVideoBrowserDownloadScreenShot.jpg)
+![Doownload Queue]({{site.image_folder}}{{site.version}}/CekliVideoBrowserDownloadScreenShot.jpg)
 Download queue
 
 The url text input field can be used for youtube search if we type in non url format.
 
-![Url search mode]({{site.image_folder}}/{{site.version}}/CekliVideoBrowserYoutubeSearchScreenShot.jpg)
+![Url search mode]({{site.image_folder}}{{site.version}}/CekliVideoBrowserYoutubeSearchScreenShot.jpg)
 Url text input as youtube search
 
 The browser or dialog tabs can be move to other position or to a new window
 
-![Url search mode]({{site.image_folder}}/{{site.version}}/CekliVideoBrowserMoveBrowserTabToNewWIndow.gif)
+![Url search mode]({{site.image_folder}}{{site.version}}/CekliVideoBrowserMoveBrowserTabToNewWIndow.gif)
 By dragging the tab header, the position of the tab can be changed
 
 Context menu using right mouse click for copy, paste, open in new Window or tab and print.
-![Url search mode]({{site.image_folder}}/{{site.version}}/CekliVideoBrowserContextMenu.jpg)
+![Url search mode]({{site.image_folder}}{{site.version}}/CekliVideoBrowserContextMenu.jpg)
 Context menu for additional functionality
 
 **Support this project**
